+++ .github/CHANGELOG.md
@@ -11,7 +11,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
  - Added an editor-only window to quickly retrieve an OAuth access token (menu : `Blizzard API/Get Access Token`)
    - This is mostly useful in Blizzard's documentation, by providing an access token to test endpoints
   
-## [0.39.0] - 2020-04-xx
+## [0.40.0] - 2020-05-01
+### Added
+ - Added support for the following WoW Profile endpoint category :
+   - Character Hunter Pets
+  
+## [0.39.0] - 2020-05-01
 ### Added
  - Added support for the following WoW Profile endpoint category :
    - Character Equipment